+++ NEWS.md
@@ -1,5 +1,13 @@
 # parsnip (development version)
 
+## Other Changes
+
+ * `tidyr` >= 1.0.0 is now required. 
+ 
+## New Features
+
+ * A new main argument was added to `boost_tree()` called `stop_iter` for early stopping. The `xgb_train()` function gained arguments for early stopping and a percentage of data to leave out for a validation set. 
+ 
 # parsnip 0.1.1
 
 ## New Features